refactor: create class & move method
